partial_json_validation/requirements.txt
anthropic==0.45.0
pydantic==2.10.4
pillow==11.0.0
